+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Releases
 
+### Unreleased
+#### New features
+  - Support for metadata which uses alg/digest instead of name/value for pins.
+    Metadata which uses name/value is still supported.
+
 ### v2.2.0 (2019-12-10)
 #### New features
   - Functionality for rebuilding the cache file (e.g. if lost) (#10)