+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Support for metadata which uses alg/digest instead of name/value for pins.
     Metadata which uses name/value is still supported.
   - Umask is used (when available) to limit access to the cache file (#79)
+  - Support for regular expressions in the switch construct (#78)
 
 #### Bugfixes
   - Removed broken support for multiple config files (#65)